surprise commit
need to do:

boss' still scaled and attack is offsetted.

crush handling: doesnt crush the server just lags it, good enough for now.

music isnt implemented.

player healthbar

decor

and tehn we ar e fucking done finally owoo odasdh

+++ 8-bit Dungeon - INternet Explorers/Assets/Scripts/CameraPlayerTrack.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 public class CameraPlayerTrack : MonoBehaviour
 {
     private GameObject player;
-    public int max_x;
-    public int max_y;
-    public int min_y;
-    public int min_x;
+    public float max_x;
+    public float max_y;
+    public float min_y;
+    public float min_x;
     // Start is called before the first frame update
     void Start()
     {
